+++ tests/Paket.Tests/Lockfile/GeneratorSpecs.fs
@@ -70,7 +70,7 @@ github "owner:project2:commit2" "folder/file.fs" """
         | Some commit ->  { Commit = commit
                             Owner = f.Owner
                             Project = f.Project
-                            Dependencies = []
+                            Dependencies = Set.empty
                             Name = f.Name } : ModuleResolver.ResolvedSourceFile
         | _ -> failwith "error")
     |> LockFileSerializer.serializeSourceFiles